feat: move StorageManager
